/balance

+++ internal/accrual/models.go
@@ -11,7 +11,7 @@ const (
 )
 
 type OrderInfoResponse struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int64  `json:"accrual,omitempty"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual,omitempty"`
 }
